+++ R/cff_gha_update.R
@@ -23,6 +23,8 @@
 #' cff_gha_update()
 #' }
 #' @export
+#'
+#' @family git
 cff_gha_update <- function(path = ".",
                            overwrite = FALSE) {
   destdir <- file.path(path, ".github", "workflows")